# api/openapi.yaml
openapi: 3.1.0
info:
  title: Example Server API
  version: 1.0.0
  description: |
    This API allows you to interact with the Example Server.

    **Global Middlewares:**
    - Logging: All requests are logged with request/response details.
...
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - {}
paths:
...
  /products:
    get:
      summary: Retrieve a list of products
      description: |
        Retrieves a list of all products in the system

        **Middlewares applied:**
        - Authentication required
      operationID: listProducts
      tags:
        - Products
      security:
        - bearerAuth: []
      Parameters:
        - name: limit
          in: query
          description: Maximum number of products to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Successfully retrieved list of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
                    description: Total number of products available
                  limit:
                    type: integer
          '401':
            description: Unauthorized - Invalid or missing authentication token
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
          '500':
            description: Internal server error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
  /products/{productId}:
    get:
      summary: Get product by ID
      description: |
        Retrieves a single product by its unique identifier

        **Middlewares applied:**
        - Authentication required
      operationId: getProductById
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid product ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /user/{userId}:
    get:
      summary: Retrieve user information and associated orders
      description: |
        Retrieves user details along with their associated orders

        **Middlewares applied:**
        - Authentication required
      operationID: getUserWithOrders
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved user information and orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOrders'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a user by ID
      description: |
        Deletes a user from the system by their unique identifier

        **Middlewares applied:**
        - Authentication required
      operationID: deleteUserById
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          description: Unique identifier of the user to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successfully deleted user
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /user:
    post:
      summary: Create a new user
      description: |
        Creates a new user in the system

        **Middlewares applied:**
        - Authentication required
      operationID: createUser
      tags:
        - Users
      requestBody:
        description: User details
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - firstname
                - lastname
              properties:
                username:
                  type: string
                  description: Username of the user
                  minLength: 3
                  maxLength: 30
                email:
                  type: string
                  format: email
                  description: Email address of the user
                firstname:
                  type: string
                  description: First name of the user
                lastname:
                  type: string
                  description: Last name of the user
      responses:
        '201':
          description: Successfully created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication. Include the token in the Authorization header as "Bearer {token}"
  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - price
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the product
        name:
          type: string
          description: Name of the product
          minLength: 1
          maxLength: 100
        description:
          type: string
          description: Detailed description of the product
        price:
          type: number
          format: float
          description: Price of the product
          minimum: 0
        category:
          type: string
          description: Category of the product
        inStock:
          type: boolean
          description: Availability of the product
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the product was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the product was last updated
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
          example: "PRODUCT_NOT_FOUND"
        message:
          type: string
          description: Human-readable error message
          example: "The requested product could not be found"
        details:
          type: string
          description: Additional error details
    Order:
      type: object
      required:
        - id
        - productId
        - quantity
        - totalPrice
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the order
        # The products schema should contain an array of objects, the object should be the product ID and the quantity ordered
        products:
          type: array
          description: List of products in the order with their quantities
          items:
            type: object
            required:
              - productId
              - quantity
            properties:
              productId:
                type: string
                format: uuid
                description: Unique identifier for the product
              quantity:
                type: integer
                description: Quantity of the product ordered
                minimum: 1
        totalPrice:
          type: number
          format: float
          description: Total price for the order
          minimum: 0
        accruedLoyaltyPoints:
          type: integer
          description: Loyalty points accrued from this order
          minimum: 0
        orderDate:
          type: string
          format: date-time
          description: Timestamp when the order was placed
        status:
          type: string
          description: Current status of the order
          enum:
            - PENDING
            - PROCESSING
            - SHIPPED
            - DELIVERED
            - CANCELED
    User:
      type: object
      required:
        - id
        - username
        - email
        - firstname
        - lastname
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user
        username:
          type: string
          description: Username of the user
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
          description: Email address of the user
        firstname:
          type: string
          description: First name of the user
        lastname:
          type: string
          description: Last name of the user
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the user was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the user was last updated
        loyaltyPoints:
          type: integer
          description: Loyalty points accumulated by the user
          minimum: 0
    UserOrders:
      type: object
      required:
        - user
        - orders
      properties:
        user:
          $ref: '#/components/schemas/User'
        orders:
          type: array
          description: List of orders associated with the user
          items:
            $ref: '#/components/schemas/Order'